+++ loading_screen.py
@@ -1,0 +1,156 @@
+import time
+import os
+
+from utilities import *
+
+class LoadingScreen():
+
+    scrole = [''' _____  _____ 
+( ___ )( ___ )
+ |   |  |   | 
+ |   |  |   | 
+ |___|  |___| 
+(_____)(_____)''',''' _____         _____ 
+( ___ )-------( ___ )
+ |   |         |   | 
+ |   |   ____  |   | 
+ |   |  / ___| |   | 
+ |   | | |     |   | 
+ |   | | |___  |   | 
+ |   |  \\____| |   | 
+ |___|         |___| 
+(_____)-------(_____)''',''' _____              _____ 
+( ___ )------------( ___ )
+ |   |              |   | 
+ |   |   ____       |   | 
+ |   |  / ___|__ _  |   | 
+ |   | | |   / _` | |   | 
+ |   | | |__| (_| | |   | 
+ |   |  \\____\\__,_| |   | 
+ |___|              |___| 
+(_____)------------(_____)''',''' _____                  _____ 
+( ___ )----------------( ___ )
+ |   |                  |   | 
+ |   |   ____      _    |   | 
+ |   |  / ___|__ _| |_  |   | 
+ |   | | |   / _` | __| |   | 
+ |   | | |__| (_| | |_  |   | 
+ |   |  \\____\\__,_|\\__| |   | 
+ |___|                  |___| 
+(_____)----------------(_____)''',''' _____                   _____ 
+( ___ )-----------------( ___ )
+ |   |                   |   | 
+ |   |   ____      _     |   | 
+ |   |  / ___|__ _| |_   |   | 
+ |   | | |   / _` | __|  |   | 
+ |   | | |__| (_| | |_   |   | 
+ |   |  \\____\\__,_|\\__|  |   | 
+ |___|                   |___| 
+(_____)-----------------(_____)''',''' _____                            _____ 
+( ___ )--------------------------( ___ )
+ |   |                            |   | 
+ |   |   ____      _        _     |   | 
+ |   |  / ___|__ _| |_     / \\    |   | 
+ |   | | |   / _` | __|   / _ \\   |   | 
+ |   | | |__| (_| | |_   / ___ \\  |   | 
+ |   |  \\____\\__,_|\\__| /_/   \\_\\ |   | 
+ |___|                            |___| 
+(_____)--------------------------(_____)''',''' _____                                 _____ 
+( ___ )-------------------------------( ___ )
+ |   |                                 |   | 
+ |   |   ____      _        _          |   | 
+ |   |  / ___|__ _| |_     / \\   _ __  |   | 
+ |   | | |   / _` | __|   / _ \\ | '__| |   | 
+ |   | | |__| (_| | |_   / ___ \\| |    |   | 
+ |   |  \\____\\__,_|\\__| /_/   \\_\\_|    |   | 
+ |___|                                 |___| 
+(_____)-------------------------------(_____)''',''' _____                                     _____ 
+( ___ )-----------------------------------( ___ )
+ |   |                                     |   | 
+ |   |   ____      _        _         _    |   | 
+ |   |  / ___|__ _| |_     / \\   _ __| |_  |   | 
+ |   | | |   / _` | __|   / _ \\ | '__| __| |   | 
+ |   | | |__| (_| | |_   / ___ \\| |  | |_  |   | 
+ |   |  \\____\\__,_|\\__| /_/   \\_\\_|   \\__| |   | 
+ |___|                                     |___| 
+(_____)-----------------------------------(_____)''',''' _____                                      _____ 
+( ___ )------------------------------------( ___ )
+ |   |                                      |   | 
+ |   |   ____      _        _         _     |   | 
+ |   |  / ___|__ _| |_     / \\   _ __| |_   |   | 
+ |   | | |   / _` | __|   / _ \\ | '__| __|  |   | 
+ |   | | |__| (_| | |_   / ___ \\| |  | |_   |   | 
+ |   |  \\____\\__,_|\\__| /_/   \\_\\_|   \\__|  |   | 
+ |___|                                      |___| 
+(_____)------------------------------------(_____)''',''' _____                                             _____ 
+( ___ )-------------------------------------------( ___ )
+ |   |                                             |   | 
+ |   |   ____      _        _         _      ___   |   | 
+ |   |  / ___|__ _| |_     / \\   _ __| |_   / _ \\  |   | 
+ |   | | |   / _` | __|   / _ \\ | '__| __| | | | | |   | 
+ |   | | |__| (_| | |_   / ___ \\| |  | |_  | |_| | |   | 
+ |   |  \\____\\__,_|\\__| /_/   \\_\\_|   \\__|  \\___/  |   | 
+ |___|                                             |___| 
+(_____)-------------------------------------------(_____)''',''' _____                                                 _____ 
+( ___ )-----------------------------------------------( ___ )
+ |   |                                                 |   | 
+ |   |   ____      _        _         _      ___   __  |   | 
+ |   |  / ___|__ _| |_     / \\   _ __| |_   / _ \\ / _| |   | 
+ |   | | |   / _` | __|   / _ \\ | '__| __| | | | | |_  |   | 
+ |   | | |__| (_| | |_   / ___ \\| |  | |_  | |_| |  _| |   | 
+ |   |  \\____\\__,_|\\__| /_/   \\_\\_|   \\__|  \\___/|_|   |   | 
+ |___|                                                 |___| 
+(_____)-----------------------------------------------(_____)''',''' _____                                                     _____ 
+( ___ )---------------------------------------------------( ___ )
+ |   |                                                     |   | 
+ |   |   ____      _        _         _      ___   __  __  |   | 
+ |   |  / ___|__ _| |_     / \\   _ __| |_   / _ \\ / _|/ _| |   | 
+ |   | | |   / _` | __|   / _ \\ | '__| __| | | | | |_| |_  |   | 
+ |   | | |__| (_| | |_   / ___ \\| |  | |_  | |_| |  _|  _| |   | 
+ |   |  \\____\\__,_|\\__| /_/   \\_\\_|   \\__|  \\___/|_| |_|   |   | 
+ |___|                                                     |___| 
+(_____)---------------------------------------------------(_____)''',''' _____                                                       _____ 
+( ___ )-----------------------------------------------------( ___ )
+ |   |                                                       |   | 
+ |   |   ____      _        _         _      ___   __  __ _  |   | 
+ |   |  / ___|__ _| |_     / \\   _ __| |_   / _ \\ / _|/ _| | |   | 
+ |   | | |   / _` | __|   / _ \\ | '__| __| | | | | |_| |_| | |   | 
+ |   | | |__| (_| | |_   / ___ \\| |  | |_  | |_| |  _|  _| | |   | 
+ |   |  \\____\\__,_|\\__| /_/   \\_\\_|   \\__|  \\___/|_| |_| |_| |   | 
+ |___|                                                       |___| 
+(_____)-----------------------------------------------------(_____)''',''' _____                                                         _____ 
+( ___ )-------------------------------------------------------( ___ )
+ |   |                                                         |   | 
+ |   |   ____      _        _         _      ___   __  __ _ _  |   | 
+ |   |  / ___|__ _| |_     / \\   _ __| |_   / _ \\ / _|/ _| (_) |   | 
+ |   | | |   / _` | __|   / _ \\ | '__| __| | | | | |_| |_| | | |   | 
+ |   | | |__| (_| | |_   / ___ \\| |  | |_  | |_| |  _|  _| | | |   | 
+ |   |  \\____\\__,_|\\__| /_/   \\_\\_|   \\__|  \\___/|_| |_| |_|_| |   | 
+ |___|                                                         |___| 
+(_____)-------------------------------------------------------(_____)''',''' _____                                                               _____ 
+( ___ )-------------------------------------------------------------( ___ )
+ |   |                                                               |   | 
+ |   |   ____      _        _         _      ___   __  __ _ _        |   | 
+ |   |  / ___|__ _| |_     / \\   _ __| |_   / _ \\ / _|/ _| (_)_ __   |   | 
+ |   | | |   / _` | __|   / _ \\ | '__| __| | | | | |_| |_| | | '_ \\  |   | 
+ |   | | |__| (_| | |_   / ___ \\| |  | |_  | |_| |  _|  _| | | | | | |   | 
+ |   |  \\____\\__,_|\\__| /_/   \\_\\_|   \\__|  \\___/|_| |_| |_|_|_| |_| |   | 
+ |___|                                                               |___| 
+(_____)-------------------------------------------------------------(_____)''',''' _____                                                                    _____ 
+( ___ )------------------------------------------------------------------( ___ )
+ |   |                                                                    |   | 
+ |   |   ____      _        _         _      ___   __  __ _ _             |   | 
+ |   |  / ___|__ _| |_     / \\   _ __| |_   / _ \\ / _|/ _| (_)_ __   ___  |   | 
+ |   | | |   / _` | __|   / _ \\ | '__| __| | | | | |_| |_| | | '_ \\ / _ \\ |   | 
+ |   | | |__| (_| | |_   / ___ \\| |  | |_  | |_| |  _|  _| | | | | |  __/ |   | 
+ |   |  \\____\\__,_|\\__| /_/   \\_\\_|   \\__|  \\___/|_| |_| |_|_|_| |_|\\___| |   | 
+ |___|                                                                    |___| 
+(_____)------------------------------------------------------------------(_____)''']
+
+    def paint():
+         write("=== Welcome to ===")
+         for frame in LoadingScreen.scrole:
+            os.system('cls')
+            print("=== Welcome to ===")
+            print(frame)
+            time.sleep(0.15)
